+++ app/jobs/fan_out_printer_photo_captures_job.rb
@@ -1,0 +1,11 @@
+class FanOutPrinterPhotoCapturesJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    Printer.find_each do |printer|
+      next unless printer.camera? || printer.prusalink?
+
+      PrinterPhotoCaptureJob.perform_later(printer.id)
+    end
+  end
+end